+++ test/fixtures/multifile/withdotimport/main.go
@@ -1,0 +1,10 @@
+package main
+
+import . "./mathpkg"
+
+func main() {
+    x := Add(10, 5)
+    y := Square(4)
+    println(x)
+    println(y)
+}